% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prueba.R
\name{hello_world}
\alias{hello_world}
\title{hello_world}
\usage{
hello_world(name = "world")
}
\arguments{
\item{name}{a name for the greeting}
}
\value{
string a greeting
}
\description{
The typical hello world example.
}

